Moving browser under platform directory Fixing Rollup config
