--- packages/realm/src/browser/tsconfig.json
@@ -1,19 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-      "noResolve": false,
-      "incremental": true,
-      "types": [],
-      "lib": [
-        "DOM",
-        "ES2022"
-      ]
-    },
-    "exclude": [
-      "../tests/",
-      "../node/",
-      "../react-native/",
-      "../../type-tests/"
-    ],
-  }
-
