+++ packages/realm/src/platform/node/tsconfig.json
@@ -14,6 +14,6 @@
     "../../tests/",
     "../react-native/",
     "../../../type-tests/",
-    "../../browser"
+    "../browser"
   ],
 }
