+++ packages/realm/src/platform/react-native/tsconfig.json
@@ -16,6 +16,7 @@
   "exclude": [
     "../../tests/",
     "../node/",
-    "../../../type-tests/"
+    "../../../type-tests/",
+    "../browser"
   ],
 }
